work

+++ A_Lucky_Division.cpp
@@ -4,8 +4,8 @@ using namespace std;
 //Speed
 #define fastio() ios_base::sync_with_stdio(false), cin.tie(nullptr), cout.tie(nullptr);
 
-bool checkLucky(int n){
-    int remain;
+bool checkLucky(long long n){
+    long long remain;
     while(n!=0){
         remain = n%10;
         if(remain!=4 && remain!=7){
@@ -19,10 +19,10 @@ bool checkLucky(int n){
 
 int32_t main()
 {
-    int n;
+    long long n;
     cin>>n;
     bool ans=0;
-    for(int i=1; i<=n; i++){
+    for(long long i=1; i<=n; i++){
         if(checkLucky(i)){
             if(n%i==0){
                 ans=1;
